+++ src/templates/bespoke/bespoke.ts
@@ -2,6 +2,7 @@ import bespoke from 'bespoke'
 import bespokeForms from 'bespoke-forms'
 import bespokeClasses from './classes'
 import bespokeInactive from './inactive'
+import bespokeFragments from './fragments'
 import bespokeFullscreen from './fullscreen'
 import bespokeHash from './hash'
 import bespokeNavigation from './navigation'
@@ -20,5 +21,6 @@ export default function() {
     bespokeProgress,
     bespokeTouch(),
     bespokeOSC(),
+    bespokeFragments,
   ])
 }